[component][fuzzing] Consolidate engine main functions
This CL extracts the common parts of the realmfuzzer and test engines
into a function templated on how to create the runner. Notably, it does
not update the libfuzzer engine: the code that integrates its engine,
runner and controller provider is missing! That omission will be
addressed in a subsequent CL.

Change-Id: I20a8fe357430ab194bf9c59bcfad6670700fb7ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695590
Commit-Queue: Aaron Green <aarongreen@google.com>
Reviewed-by: Dmitry Yatsushkevich <dmitryya@google.com>