[kernel][thread] Fix timeslice extension bug
This commit fixes a bug with timeslice extensions where a thread with
a timeslice extension may become preempted prior to even activating
its extension.

The logic flaw is that once the program saw the preempt count was zero
and that there was no active extension it called FlushPending and
relied on there being no local preempts_pending_.  The flawed
reasoning was that preempts_pending_ can't possibly contain the local
CPU at this point because we just observed that there's no active
extension and setting a local preempts_pending_ will always activate
any inactive extension.  The problem with this reasoning is that it
does not take into account the fact that a call to PreemptReenable may
be interrupted and set a local preempts_pending_ after the check for
an active extension, but before the call to FlushPending.

The commit updates PreemptReenable, PreemptReenableDelayFlush, and
EagerReschedReenable, adds DEBUG_ASSERTs, and in-kernel unit tests.

Rename ExpireActiveTimesliceExtension to
ClearActiveTimesliceExtensionIfExpired to improve clarity.

Bug: 104468
Change-Id: I7a6d704af2f216c920d4cba2eabe439679f2e391
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704422
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>