[docs] Document the big-picture for the configuration of components.
This is the first publicly available version of a family of documents
that we've developed and discussed internally for a while.

The structure of how this integrates with docs is based on a very
useful discussion with Nick - thanks! Adding the reviewers he suggested
along with the people who reviewed in its most recent doc form.

Fixed: 95374
Change-Id: Ia90f307c3a8fc70604f0a392d05d286bdd89d6dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701246
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Guinevere de la Mare <guinevere@google.com>