[fidl][server_suite] Top-level structs, tables, unions
This CL adds tests for top-level structs, tables and unions in requests
and responses.

For tables, where there is optionality, bindings may or may not check
for the presence of the value. It is not necessarily important to
require checking for presence / absence in the binding because the
test runner will ultimately check the correct presence bytes are
received, so no mandate will be issued for untested behavior like
this.

Bug: 102339
Change-Id: I855bef12552fe2564e17b79f774deb0801f142ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704014
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>