[kms] Remove the CFv1 component from the build
KMS is not load-bearing in its current form, and is currently blocking
the CFv1 -> CFv2 migration of `tee_manager`.  Removing the component
definition causes no functional or coverage regression, but does help
unblock the v2 migration for other components.

We keep the code (and unit test) in-tree to be referenced and
potentially reused for TPM-related work and can make a decision whether
to bring a `kms` component back or remove it later.

Bug: 103105
Change-Id: I47b689d991a176b78eb66266a8a8966b23f2f0e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704407
Fuchsia-Auto-Submit: Drew Fisher <zarvox@google.com>
Reviewed-by: Jody Sankey <jsankey@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>
Reviewed-by: Ben Wright <benwright@google.com>