[dfv2] Delete OwnedOffer
Now that we do not re-generate offers during the lifetime
of a node, we can remove this struct that stores an owned
offer.

If we do need to make a deep copy of a fidl table, convert
it from wire format to natural format and then back.

Change-Id: I7d52d43c8dea9287b60de6d2d3145c03b31f5a3f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704048
Reviewed-by: Sarah Chan <spqchan@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>