[fidlgen] Namespace HandleSubtype values
The current values to the HandleSubtype 'enum' are not namespaced and
are exported with ambiguous names like `fidlgen.Channel`. It is clearer
to prefix these with the enum name ("HandleSubtype"). This change also
eases the upcoming introduction of a type to mirror the IR for resource
declarations (a more natural fit to `fidlgen.Resource`).

We also s/HandleSubtypeHandle/HandleSubtypeNone, as the latter is
clearer and more precise.

Bug: 7660
Change-Id: I67eef76f22ae07c998638df60af8243f91177b3f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704242
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>