[ftl] Diagnosis tool to forward ECC failure information to FTL.
Instead of skipping the failed ECC pages and leaving them blank, set the
ECC failure status in the ndm so that it will fail when the FTL reads
it. This will more closely emulate the behaviour on the prod device.

Also "continues" on failed reads instead of breaking while diagnosing
PBR, since read failures are actually expected to happen sometimes now
and have nothing to do with PBR.

Bug: b/240347573
Change-Id: I225e288ef9debeef007431f4c42e37cea550be29
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704663
Reviewed-by: Brandon Castellano <bcastell@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>