[pwrbtn-monitor] Fix test flake
Fix issue where test would flake if the first async routine to
connect to the power statecontrol service was the OOM routine.
When this happened the disconnection of the routine caused the
oneshot channel to be dropped, making the test complete too early
and leading to errors.

MULTIPLY: pwrbtn-monitor-integration-test: 500

Fixed: 103485
Change-Id: I1d04bb79c3cbf1d1c7ed418410f8d61ca05c074b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704537
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Justin Mattson <jmatt@google.com>