[ffx] Improve network validation
This CL adds some new checks to network validation for Qemu-based
emulators. Specifically, the Tap checks remain unchanged; but if the
mode is Auto, we bail (since mode resolution is already complete); if
User, nothing; and if None, we
require --console or --monitor, since there will be no other way to
communicate with the emulator.

Bug: 101313

Change-Id: Ic1f6d06fa2056510d76055321501cef19b45d08d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/703875
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>