[fidl][cpp] Client support for fpromise promises
Today fpromise is the C++ promise library used in production. This CL
adds a helper library to transform callback-style async FIDL calls to
promises.

Because there is a new C++ fasync futures library on the horizon, this
CL opts to keep fpromise support in a separate library instead of built
into the C++ bindings. We'd like to reserve the built-in support for
fasync.

Change-Id: I5b0076c998209d39a5dbabd7a5c292128bb11352
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698173
Reviewed-by: Zachary Stewart <zstewart@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>