[netstack3] Add connect_udp_listener function
Add a function to replace an existing UDP listener with a connected
socket. This makes it possible for Bindings to implement POSIX
connect() as an atomic operation in Core. Making this an atomic
operation makes it simple for Core to propagate socket options when
connect() succeeds.

Bug: 96573
Change-Id: I6272576ebbc30bc0802a85b2914f7b30431ae3b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662314
Reviewed-by: Zeling Feng <zeling@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>