[test-ui-stack] Implement simple touch input helper
This change implements a very simple touch input helper component, which
will serve as the basis for the v2 input synthesis component.

NOTE: Some of the code here is copied from the v1 input-synthesis library.
We are forking that code consciously, as it will be the basis for OOT
input testing and SL4F.

Test: fx test input-helper-tests, fx test flutter-embedder-test-ip with
fxr/699870
Bug: 104803

Change-Id: I4805b5e8fd541ab4c3e605f0a84ba20009f4aa2c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/702985
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Commit-Queue: Alexander Brusher <abrusher@google.com>