[elfldltl] Add <lib/elfldltl/load.h> template API
This starts the support facilities for loader implementations.
To begin with, it just provides a few template APIs that wire up
the File and Diagnostics APIs with layout.h and phdr.h support
code to cover the first stages for reading headers for loading.

Bug: 91206
Change-Id: Ic075433248001d7f49a4b7ae2f5b31ef0c186675
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/629043
Reviewed-by: Haowei Wu <haowei@google.com>
Reviewed-by: Alex Brachet <abrachet@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>