[lowpan][openthread_rust]: Fix for UDP `ot::Message` leak.
It would appear that a previous assumption---that the UDP handler
method would take ownership of the underlying otMessage instance---was
incorrect. This is leading to `otMessage` instances leaking.

This change makes sure that we don't attempt to pass ownership, which
ensures that the `ot::Message` instance is properly cleaned up.

Bug: 105401
Change-Id: I4847b2979b0dd2df44cc7131ec2a407eba88211e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705268
Commit-Queue: Jiaming (Charlie) Wang <jiamingw@google.com>
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>