[gfx] Add Color Conversion API
Hookup the new color conversion API functions to GFX via the Engine. This allows clients of the ColorConversion FIDL API to call
SetColorConversionValues and SetMinimumRGB on the Scenic GFX backend.

The old method of using the GfxCommandApplier still exists in the code, and will be removed once this API is hooked up to SceneManager.

Fixed: 96971

Change-Id: I272e96f6a96461328a061a13844438d2c55525e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704028
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: Chris Fontas <cfontas@google.com>