[component][fuzzing] Make generic fuzzing integration tests
This CL moves most of the realmfuzzer integration tests under
//src/sys/fuzzing/common, in much the same manner that most of the
Runner unit tests are under common/ and use macros to specialize for the
specific runners. This is done for the same purpose as well: to
facilitate creating integration tests that are the same for both
libfuzzers and realmfuzzers.

Change-Id: Iccdd7ae577bcf2196e05c92928d324db8268cd7a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695591
Commit-Queue: Aaron Green <aarongreen@google.com>
Reviewed-by: Dmitry Yatsushkevich <dmitryya@google.com>