[fidl] Client-side dynamic test suite
This CL introduces a client side dynamic test suite to FIDL, which is
meant to test edge cases of FIDL client bindings. It follows similar
patterns as the server side test suite, but tests are driven from
clients in bindings under test, rather than the harness.

A VERIFY_* mechanism is introduced that parallels ASSERT_*, but reports
errors to the test runner.

The Channel and Bytes classes are factored out of the server test suite
to a shared location, which meant not using test assertions directly in
that logic.

Bug: 102339
Change-Id: I4668848e895862924744b684464f74b6d3da9d0e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700344
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>