[starnix] Correctly preserve execute permissions in fork
The bug: If the same vmo is mapped twice, once using a handle without the execute right and once with, and the mapping traversal found the
non-executable mapping first, the hashmap cache would get a clone of the non-executable handle and this would be reused for the mapping that was supposed to be executable.

The fix: delete the hashmap. It was intended as an optimization to avoid making extra clones, but it could never have done that in the first place, because you can't map a private VMO twice. The only entries retrieved from the hashtable would be shared VMOs, and you could have just cloned the handle instead of looking it up.

Change-Id: I31ef84e8f4ee4fe258d1e5a04ff4970052ea0c0b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701604
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Theodore Dubois <tbodt@google.com>