[scripts] Remove -S flag
This flag isn't available for the `env` provided by GNU coreutils 8.28
and will return an "invalid option -- S" error in this invocation.
This flag can be removed because (1) it's intended usage of splitting
an argument string isn't actually be used here and (2) it's incorrectly
used here because it's not being passed a string argument.

Bug: 62896
Change-Id: I6a0513a29c4839502ccb51ec8a029b2315fc57bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705624
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>