[codec] stress decode of h264 and vp9 + PCMM
This is mainly a repro and regression test for bug
b/236163568.

This runs 3 threads, which randomly select decoding an h264
stream decoding into protected buffers, a vp9 stream decoding
into protected buffers, or making some random-size allocations
and deallocations using protected_memory_size (on devices that
have AMLOGIC_SECURE heap).

The combination of these happening concurrently was enough to
catch the problem where writing to wid + 0xc register for
_any_ HW protection range (4-14) in the BL32 firmware can
occasionally glitch the H264 decoder (even if writing back the
same value, and for no obvious reason).

Previous to this CL, which adds this test to run on relevant
HW, we rolled a firmware change that avoids this problem with
steady-state writes to wid + 0xc by not writing to wid + 0xc
(outside of setup steps prior to any H264 decoder use), and
instead dedicates one of the HW protection ranges to zeroing
during un-protect (range 4).  So with that change, this test is
expected to pass reliably (including with the new concurrent
PCMM stress, now that the bug is fixed).

MULTIPLY: use_h264_and_vp9_decoders_and_pcmm_stress_test: 60
BUG: b/236163568
Change-Id: I74054998f5f40729faca7c210f5a7dca785e55e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/697582
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>