[f2fs] Fix a corruption bug on a Page bitmap
The CL fixes a bug in VmoNode::CreateAndLockVmo() where an wrong
offset was used to calculate the range of Pages that kernel had
reclaimed in a VmoNode. It would result in a corrupted Page or
a page fault due to corrupting or overflowing on VmoNode::page_bitmap_
keeping the valid Page information for each VmoNode.

Multiply: f2fs-unittest: 200
Bug: 105351

Change-Id: I02df827236ba0cd5a977a62f93118a1f2437aa08
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706222
Commit-Queue: Seokhwan Kim <sukka.kim@samsung.com>
Reviewed-by: Aaron Drew <ripper@google.com>