[fdf] Add Device Group Manager to driver manager
Add classes for DeviceGroupManager, DeviceGroup, and
CompositeManagerBridge.

The DeviceGroupManager is responsible for managing
device groups.

The DeviceGroup is partially abstract class for
managing the state of a device group.

The CompositeManagerBridge is a bridge interface that
allows DeviceGroupManager to agnostically interact
with the DFv1 and DFv2 implementations for composite
handling.

Bug: 103182
Change-Id: If8b1215320b0d1158997ee95dae85b9e4f462868
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/703684
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Sarah Chan <spqchan@google.com>