[hid-parser] Limit the size of device descriptor
There is such a thing as allocating too many bytes for device
descriptors. The fact that we can allocate 2GB of memory for this
purpose doesn't mean that we should.

This change puts an upper limit to the allocation size to 1GB, I hope
that this would be enough to handle realistic usage.

Bug: 64791
Change-Id: Icb2d0d5bf7e96e4083249153f12e59caa9dcf86b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705484
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>