[ffx][pkg] Add transparent gzip support to GcsRepository
When a file is gzipped compressed on GCS, GCS is unable to report the total
content length. To fix this, this changes GcsRepository to cache the file
into a temporary location so we can compute the total length, then return
the stream to the caller.

Fixed: 105309
Change-Id: I56df0672087545ad738e449bb2c026e29464f824
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706103
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: James Buckland <jbuckland@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>