[kernel][vm] Free failed pages in PhysicalPageProvider
If SupplyPages() fails, the pages left over in the source splice list
need to be freed. Fix the freeing logic which was previously freeing
from the page list that the splice list was created from, and would
therefore be empty after the move.

Run-All-Tests: true
Change-Id: I7b1f185e0f05771645daf9ac5a7c08f7d267b974
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705625
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>