[intel-hda] Fix device_get_fragment_protocol
The last argument of this function is a void* that must
be the protocol struct. Passing a ddk:Pci here is not
allowed, even though it works because it wraps the
protocol.

Fix this by using the ddk::Pci connect function.

Change-Id: Ieaa87115eb35d8a92076e8ca10fa232d850fe6f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706309
Reviewed-by: Ian Fisher <iafisher@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>