[ntr] Add option to wait for any IP address
Some tests want to be able to wait until the interfaces they add to the
hermetic network realm have their autoconfigured IP addresses and have
completed duplicate address detection, so we add an option to
add-interface to wait for any IP address to be assigned to the interface
before returning. This is not strictly guaranteed to be the
autoconfigured IP address, but it should be unless the client is
intentionally assigning an IP address statically by separate means.

Bug: 105295, 105506
Multiply: network-test-realm-integration-tests
Change-Id: Ie2948242c4470cb55d43449ff8294b22b90ec638
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705542
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>