[test_manager] Refactor to introduce new scheduler
This change is a refactor of test_manager that allows for the
introduction of new test suite scheduling strategies. This change
introduces a new scheduler module that contains the Scheduler trait.
This trait can be implemented by any object and dictates how a
collection of suites is scheduled and executed. This change also adds a
new enum to the facets module that indicates whether a facet is resolved
or unresolved. This will be useful when we add the ParallelScheduler
because the ParallelScheudler will need to resolve test suite facets
before invoking run_single_suite. We want to avoid resolving suites
twice so checking the status of the facet resolution in run_single_suite
will be necessary. Please refer to the following design doc for more
information - go/fuchsia-test-manager-refactor

Test: To test my changes, I made sure running "fx test
      test_manager_test" still passed all tests.
Change-Id: I3aa1fe7c8e55ff0d4c1fc06bacb781de5f355ae0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705599
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Ismael Castro <iicastro@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>