[brcmfmac] Fix compiler errors when kALL filter is on
Turning on the kALL filter for BRCMF_DBG no longer causes the build to
fail to compile. The real fix for this problem is to add a build
target that requires the kALL filter to be on. This way we cannot
check-in code that fails to compile when the filter is on.

This change renames len to length_ in the macro expansion so there is
less chance of a conflict with variable used in the scopes where the
macros expand. (This is actually a generic problem with macros in C.)

Test: None. Debug code only.

Change-Id: If1c004dee30d92f586a380f08668c41147a92c21
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704528
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>