[netstack3] Implement disconnect for UDP
Add support in Core and Bindings for the Disconnect FIDL call that
corresponds to POSIX connect() with AF_UNSPEC. This allows applications
to remove the remote address for a connected socket, turning it into a
listening socket.

No tests are added in this patch, but it should cause some out-of-tree
tests to start passing.

Change-Id: I1bb5f946c74c2f1e498893b86ce35816edf8ebd9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705186
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>