[mediasession][examples] Convert example_player to v2
Convert the media session example player into a V2 component that can be
monitiored with the `mediasession_cli_tool`.

An `example-players` collection is added to the
`mediasession.core_shard.cml`: this shard provides the needed
capabilities from "#mediasession" to the example player V2 component and
the fx shell binary `mediasession_cli_tool`, which is used to
control/monitor the state of one or more example players. The
mediasession core shard is used instead of a dedicated core shard for
the example player because it already contains the needed capabilities
for both the example player and cli tool.

A README is added to describe how to configure Fuchsia and run the
example player component with the `mediasession_cli_tool`.

Bug: 103877
Change-Id: I404a03f57ed891480080a73fc330717af596f7d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705035
Reviewed-by: Dale Sather <dalesat@google.com>
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>