[benchmarks] Add initial allowlist for performance metric names
This adds an initial implementation of an allowlist and applies it to
a single test.  More tests will be covered in later changes.

The process for adding and updating an expectations file is not yet
very smooth.  I will improve that in a later change too.

sl4f.api file: In practice there is no API change because these parts
of performance.dart are not used via the Fuchsia SDK.
MetricsAllowlist is only made public so that it can be tested in
performance_test.dart.

Bug: 105202
Test: fx set terminal.x64 --release
  + fx test --e2e host_x64/fidlc_microbenchmarks_test
Change-Id: I6b1147c6a7400a111393909f0d462b44f1c4aabf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705628
Reviewed-by: Fadi Meawad <fmeawad@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
API-Review: Gary Bressler <geb@google.com>