[driver_runtime] Add protocol discovery runtime APIs.
This is only the runtime part. Implementation in next CL.

To connect to a parent protocol, a child driver will create
a token ChannelPair and pass one end to the parent. The
child driver will then call fdf_token_exchange with the
remaining end of the token.

The parent driver will call fdf_token_register with the
received token and protocol on_exchange callback.

Bug: 102201
Change-Id: I8cace35270b01102ea89873632073806847ef770
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704414
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>