[build] Remove write_file() in sdk_atom().
This replaces a write_file() call with a generated_file() target
in the sdk_atom() template definition. The result are equivalent
except that the files can be generated in parallel.

Adding a new dependency to src/tests/end_to_end/fidlcat/BUILD.gn
is required to avoid GN complaining that a target uses an input
file without depending on the target that generated it.

In practice, this only touches 4 files in  a core.x64 build config:

- gen/src/zircon/lib/zircon/zircon_sdk.sdk
- host_arm64/gen/src/developer/ffx/plugins/emulator/aemu_internal.sdk
- host_arm64/gen/src/developer/ffx/plugins/emulator/qemu_internal.sdk
- x64-shared/gen/src/zircon/lib/zircon/zircon_sdk.sdk

Bug: None
Change-Id: I27dbc8047b113c81c47b9582a0b116f2239c6a2c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706022
Reviewed-by: Dangyi Liu <dangyi@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: David Turner <digit@google.com>