[netstack3] Add reconnect_udp function in Core
Add the ability to atomically update the destination of connected
sockets to Core. This is required to support operation of POSIX
connect(). While it could be implemented as remove-then-connect, pushing
the functionality into Core makes it easier to keep socket options from
one to the next.

Bug: 96573
Change-Id: I0191f807df75403cdf7dd7bff30f4fecfc58d647
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/665883
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>