[fidl][cpp] In-args: use `const T&` for value and `T` for resource
The natural messaging APIs have signatures like so today:

  Foo(fidl::Request<Foo> request);

this causes extra copying. This CL changes those to look like:

  Foo(const fidl::Request<Foo>& request);

for when the domain object is a value, and

  Foo(fidl::Request<Foo> request);

for when the domain object is a resource.

The completer APIs used looks like so today:

  void Reply(fidl::Response<Foo>&& response);

The `T&&` is very subtle:
- it forces the user to always std::move their value in
- the signatures suggests that the function consumes the response, but
  is actually not the case: nothing will happen to `response` unless
  `Reply` explicitly moves it elsewhere.

So that's changed to

  void Reply(const fidl::Response<Foo>& response);
  void Reply(fidl::Response<Foo> response);

as well in this CL.

Change-Id: I87f8ab27ab9b7cd5595a27923c50f406046fd267
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706303
Commit-Queue: Yifei Teng <yifeit@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: Zachary Stewart <zstewart@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>