[fxtest] Populate FUCHSIA_NODENAME for local host tests
Uses get-device-name to populate the FUCHSIA_NODENAME environment
variable for host tests. FUCHSIA_NODENAME is provided by infra scripts
when these tests are run in CI/CQ, but the variable is not provided
when tests are run locally.

Testing: manually verify that the nodename is available to a host test
that uses it.

Change-Id: I204057d0e6866ad503d0bc02fe765da79ce5c259
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705736
Reviewed-by: Gary Boone <gboone@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>