[ffx][package] Depend on test empty-repo
The coverage builder appears to be getting confused because of how we
were depending on the `empty-repo` artifacts from the `pkg` library. The
files are listed as a test dependency, but the `test_util` module is
exported as a non-test module. This tries to fix the test by directly
depending on these artifacts.

Bug: 105539

Change-Id: Idcc3b50e0ab45f19e0778403926420f2b00c9bcd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707027
Reviewed-by: Ben Keller <galbanum@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>