[kernel][iommu] Fix discontiguous/contiguous logic.
Previously we would attempt to see if the range we were mapping was
contiguous. If it was, we would call MapDiscontiguous(). If it wasn't, we
would call MapContiguous().

Unsurprisingly, this worked OK if the mapping _was_ contiguous, but if
it wasn't, it would fail.

Bug: 91878
Test: boot Atlas with IOMMU enabled, observe SDHCI driver works.
Change-Id: I33c4f50e800ee4dcf4623f818f784a61920226a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706885
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>