[settings] Fix flakiness in factory_reset_test
The factory reset FIDL does not return a result on the set FIDL call, so
clients are not guaranteed to know when their request has been
processed. This CL changes the mock to push the set calls it has
received to a channel so that the test can verify that the request was
processed and verify that the set call received the correct value.

Multiply: factory_reset_integration_test
Fixed: 105502
Test: fx test factory_reset_integration_test
Change-Id: I0ffc5fab703fbac671d88d910af3897d1acba1d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706902
Reviewed-by: Sijie Chen <sijiec@google.com>
Reviewed-by: Chip Fukuhara <cfukuhara@google.com>
Commit-Queue: William Xiao <wxyz@google.com>