[kernel] Simplify interrupt handlers
We can remove interrupt_eoi, as it was only needed by the hypervisor for
`zx_interrupt_bind_vcpu`. This allows us to simplify all of the
interrupt handlers in the kernel, and the interrupt handling logic in
the GIC code.

Bug: 104712
Change-Id: I8228c707d7712aca6cebc0fc1827b6894a2a9bc7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706386
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>