[fidlgen] Leverage the `Declaration` interface (and friends)
This change introduces two successive extensions to the `Declaration`
that represent layouts and "resourceable" layouts. We this in hand, we
consolidate all "loop over each declaration in Root" logic to a single `Root` method that passes each `Declaration` - on which the caller can use type assertions to derive the concrete declaration type or the other two interfaces -  and then use this to substantially tidy up the sites that rely on such iterations.

This type will be of further value in other contexts where we'd want to
deal with the totality of declarations with minimal boilerplate (e.g.,
computing the dependency graph).

Bug: 7660, 51002
Change-Id: I45bdd2d0b79765ad3c8895eac7d5395e6e5e5f91
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704243
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>