[virtualization][vsock] Wire in config provided listeners
This change wires config provided listeners in for the legacy and new
vsock device.

For the legacy device, config provided listeners are forwarded to the
host vsock. For the new device, config provided listeners are passed to
the VMM and onwards to the device.

Bug: 97355
Change-Id: I83c212fe97434bc9ca7312a8f557a3980249d11a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705405
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: David Hastings <dahastin@google.com>