[virtualization] Update linux runner to pass Listeners via config
This changes the linux runner to pass vsock Listeners that need
to be available the moment the Guest starts via config, instead of via
HostVsockEndpoint.

This is required when using the out of process vsock device which
doesn't start until the guest does.

Bug: 97355

Tested:
1) Flashed to a NUC
2) Ran "vsh -c" a bunch of times
3) Confirmed that the container loaded and we ended up on
   machina@penguin

Change-Id: Id2f0f86db23df263ee2f16165d1f8de68f751781
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706384
Commit-Queue: David Hastings <dahastin@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>