[mdns] make probe conflict checks case insensitive
This makes it so that when probe denials are sent with different case
(Example: host.local and hOsT.local) a device would recognize this as a
conflict/probe denial and choose a new hostname to probe with. Extended
current conflict tests to handle this scenario.

Bug: 105386
Test: Manually issued case insensitive probe denials and confirmed
device recognized the denial and sent a new probe with a new hostname

Change-Id: I4c9a2f05f4c9e6be60946f3fc1174bc10a2ee32b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705105
Reviewed-by: Sundarajan Srinivasan <sundarsrini@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>
Commit-Queue: Nathan Melwani <nathanmelwani@google.com>