[universe-resolver] Simplify .cmls
The base-only .cmls are no longer used, because the universe-resolver
no longer has a base-only mode, it always uses
fuchsia.pkg.PackageResolver (which in practice always comes from the
pkg-resolver) for package resolution.

The incoming fuchsia.component.resolution.Resolver capability is no
longer used (it used to be the Resolver protocol capability from the
base-resolver when universe-resolver had a base-only mode).

The for-test .cml was the same as the regular .cml.

example.cml was referencing a full-resolver component which does not
exist in the package.

Bug: 105528
Change-Id: I1180fd09da4c8e5efdae0b412774f3dc9039a594
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705684
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Fuchsia-Auto-Submit: Ben Keller <galbanum@google.com>