[netstack-gotests] Avoid deadlock on waiter events
TestTCPEndpointMapConnect is hitting a deadlock because the test is
written to assume at most 2 events are sent to the waiter constructed
for the test, but the waiter could be notified up to 3 times. The fix
is to increase the buffer size of the channel the events are being
written to.

Fixed: 97765
Multiply: netstack-gotests
Change-Id: I8dacf6010bcb5853ac5ba6edd1920a966c165d0e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706931
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>