[inet] Fix instances of UB
The core.x64-thinlto builders seem to fal on mdns tests because of UB
in the way hashes for IpAddress is done. The first type is accessing
`v6_` in `as_words` when the active member is `v4_`. The second type
would come from casting the underlying active data to a `uint16_t *`
and dereferencing that pointer if the active member is `v4_` which is
a struct that just contains a uint32_t (this breaks strict aliasing).

Not sure if this is the best fix, but this CL updates `as_words` to
only support V6 IpAddresses where the underlying type is a union that
can be a `uint16_t *`. This way, users of `as_words` can't explicitly
dereference `v4_` through an improper type. It also looks like
`as_words` is mainly relevant for V6 IPs anyway.

This also updates the hash for V4 IPs to iterate over the bytes of
`v4_` instead which is legal C++.

Fixed: 103246
Change-Id: I926c39b36f68d4bbc385bd1173c0060ec0222693
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706850
Reviewed-by: Dale Sather <dalesat@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>